abort_source: allow multiple request_abort()
When request_abort() is called more than once on abort_source,
an assertion is triggered. Since abourt_source can be aborted
from multiple sources, it is necessary to always check,
whether abort was already requested.

To simplify code calling request_abort(), the method can be called
multiple times. Each but first call is a noop.

+++ include/seastar/core/abort_source.hh
@@ -104,13 +104,14 @@ private:
     std::exception_ptr _ex;
 
     void do_request_abort(std::optional<std::exception_ptr> ex) noexcept {
-        assert(_subscriptions);
-        _ex = ex.value_or(get_default_exception());
-        auto subs = std::exchange(_subscriptions, std::nullopt);
-        while (!subs->empty()) {
-            subscription& s = subs->front();
-            s.unlink();
-            s.on_abort(ex);
+        if (_subscriptions) {
+            _ex = ex.value_or(get_default_exception());
+            auto subs = std::exchange(_subscriptions, std::nullopt);
+            while (!subs->empty()) {
+                subscription& s = subs->front();
+                s.unlink();
+                s.on_abort(ex);
+            }
         }
     }
